model-deployment/common/hydraulic/python-scoring-pipeline/scoring-pipeline/requirements.txt
numpy==1.18.1
pandas==0.24.2
git+https://github.com/h2oai/pandas-ml#egg=pandas-ml
pycryptodome==3.6.6
requests==2.20.0
setproctitle==1.1.10
filelock==3.0.12
numpy==1.18.1
pandas==0.24.2
git+https://github.com/h2oai/pandas-ml#egg=pandas-ml
requests==2.20.0
scikit-learn==0.21.2
scipy==1.3.1
setproctitle==1.1.10
toml==0.10.0
tensorflow==1.13.1
statsmodels==0.10.1
datatable-0.11.0a242-cp36-cp36m-manylinux2010_x86_64.whl
h2o4gpu-0.3.2-cp36-cp36m-linux_x86_64.whl
h2oaicore-1.8.4.1-cp36-cp36m-linux_x86_64.whl
scoring_h2oai_experiment_0abac7c4_8993_11ea_ae54_0242ac110002-1.0.0-py3-none-any.whl
